Redirect script imports based on moved source files
