+++ scripts/export-tags.js
@@ -17,9 +17,9 @@ import path from 'path';
 import { fileURLToPath } from 'url';
 import jsYaml from 'js-yaml';
 import { validateConfig } from '../src/validateConfig.js';
-import { getPlatformApiUrls } from '../src/defineApiUrls.js';
-import { getPlatformDisplay } from '../src/defineRibbonVals.js';
-import { filterNewAdditionalEntries } from '../src/filterNewAdditionalEntries.js';
+import { getPlatformApiUrls } from '../src/utils/defineApiUrls.js';
+import { getPlatformDisplay } from '../src/utils/defineRibbonVals.js';
+import { filterNewAdditionalEntries } from '../src/utils/filterNewAdditionalEntries.js';
 
 const __dirname = path.dirname(fileURLToPath(import.meta.url));
